requirements.txt
python-dotenv
selenium
bs4
requests
langchain_groq
langchain_openai
langchain